+++ src/utils/network.ts
@@ -1,23 +1,21 @@
 import * as vscode from "vscode";
+import { QUARTO_WIZARD_LOG } from "../constants";
 import { showLogsCommand } from "./log";
 
-export async function checkInternetConnection(
-	url: string = "https://github.com/",
-	log: vscode.OutputChannel
-): Promise<boolean> {
+export async function checkInternetConnection(url: string = "https://github.com/"): Promise<boolean> {
 	try {
 		const response = await fetch(url);
 		if (response.ok) {
 			return true;
 		} else {
 			const message = `No internet connection. Please check your network settings. ${showLogsCommand()}.`;
-			log.appendLine(message);
+			QUARTO_WIZARD_LOG.appendLine(message);
 			vscode.window.showErrorMessage(message);
 			return false;
 		}
 	} catch (error) {
 		const message = `No internet connection. Please check your network settings. ${showLogsCommand()}.`;
-		log.appendLine(message);
+		QUARTO_WIZARD_LOG.appendLine(message);
 		vscode.window.showErrorMessage(message);
 		return false;
 	}